新增 歌单管理功能

+++ Shared/NeteaseCloudMusicApi/NeteaseCloudMusicApi.swift
@@ -254,7 +254,7 @@ extension NeteaseCloudMusicApi {
         cancelDict["\(#function)"] = httpRequest(method: .POST, url: url, data: encrypt(text: data.json), complete: complete)
     }
     //歌单顺序
-    func playlistOrderUpdate(ids: [Int], complete: @escaping CompletionBlock) {
+    func playlistOrderUpdate(ids: [Int64], complete: @escaping CompletionBlock) {
         let url = "https://music.163.com/weapi/playlist/order/update"
         let data = [
             "ids": ids
